#5 [cxx] added C++ Support

+++ kernel/console.c
@@ -95,7 +95,7 @@ int consoleread(int user_dst, uint64 dst, int n) {
     c = cons.buf[cons.r++ % INPUT_BUF_SIZE];
 
     if (c == C('D')) { // end-of-file
-      if (n < target) {
+      if (n < (int)target) {
         // Save ^D for next time, to make sure
         // caller gets a 0-byte result.
         cons.r--;
