+++ kernel/string.c
@@ -3,7 +3,7 @@
 void* memset(void* dst, int c, uint n) {
   char* cdst = (char*)dst;
   int i;
-  for (i = 0; i < n; i++) {
+  for (i = 0; i < (int)n; i++) {
     cdst[i] = c;
   }
   return dst;
