+++ kernel/sysproc.c
@@ -41,7 +41,7 @@ uint64 sys_sleep(void) {
   argint(0, &n);
   acquire(&tickslock);
   ticks0 = ticks;
-  while (ticks - ticks0 < n) {
+  while ((int)(ticks - ticks0) < n) {
     if (killed(myproc())) {
       release(&tickslock);
       return -1;
