+++ user/ulib.c
@@ -37,7 +37,7 @@ uint strlen(const char* s) {
 
 void* memset(void* dst, int c, uint n) {
   char* cdst = (char*)dst;
-  int i;
+  uint64 i;
   for (i = 0; i < n; i++) {
     cdst[i] = c;
   }
